fix(storage): retain Episode admission facade binding
